Created new Tests in test_zipvalidator
I created 3 new tests in test_zipvalidator for our zipvalidation python file. The 3 changes were to prevent edge cases such: If someone were to rename a zip file to be different than the original directory (would throw an error before), not creating a folder could extract all files into our src folder which is messy, and finally we check to make sure we aren't overwriting a directory a user may have uploaded previously. This is a temporary fix and we will need a better fix in the future. All tests pass except the new ones and test_analyzer because there is no code for our analyzer module currently. (Delete the test analyzer file to run tests, or ignore it when running tests).