Implementation of the edge case error fixes in zip validation
I updated the unzip_file method within the zipvalidation.py module to account for the case where someone altars the name of the zipfile after creation. The extracted directory would not have the same name as the zip file and would throw an error. Created a file with the same name as the zip file before extraction and extracted all files into the folder with the same name. Also accounted for zip files that would be extracted into the src folder with the same change. Also checked to see if there was a folder already extracted with the same name in order to not overwrite previously extracted directories. 26/26 tests pass. In order to get all tests to pass, you need to delete the test_analyzer file because it tests a file with no code.