Added more testing for the resume formatter to ensure that there is 100% coverage on the file.
When we edit the resume in the future we will need to update these tests. this is because some of the tests that are being created really on text outputs in to the resume and as we develop the resume the outputs may change. We must stay on top of this